// src/config/payload.types.ts
export interface Media {
  id: string;
  type: 'PUBLIC' | 'USER';
  owner?:
    | ({
        relationTo: 'users';
        value: string | User;
      } | null)
    | ({
        relationTo: 'profiles';
        value: string | Profile;
      } | null);
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "profiles".
 */
export interface Profile {
  id: string;
  name: string;
  profilePicture?: (string | null) | Media;
  /**
   * Número de WhatsApp para contato
   */
  whatsappNumber?: string | null;
  /**
   * Configurações para o processo de exclusão de conta (LGPD)
   */
  accountDeletion?: {
    requested?: boolean | null;
    /**
     * Data em que o usuário solicitou a exclusão
     */
    requestDate?: string | null;
    /**
     * Data em que a exclusão será efetivada após o período de carência
     */
    scheduledDate?: string | null;
    reason?: string | null;
    /**
     * Período em dias antes da exclusão ser efetivada
     */
    gracePeriodDays?: number | null;
    cancelled?: boolean | null;
    cancelledDate?: string | null;
    completed?: boolean | null;
  };
  /**
   * Campo legado, use o grupo "Exclusão de Conta"
   */
  scheduledRemovalDate?: string | null;
  notificationTokens?:
    | {
        value?: string | null;
        label?: string | null;
        id?: string | null;
      }[]
    | null;
  /**
   * Campo legado, use o grupo "Exclusão de Conta"
   */
  removedAccount?: boolean | null;
  /**
   * Filtros para o perfil
   */
  filters?: string | null;
  register: {
    type: 'CRM';
    number: number;
    uf:
      | 'AC'
      | 'AL'
      | 'AP'
      | 'AM'
      | 'BA'
      | 'CE'
      | 'DF'
      | 'ES'
      | 'GO'
      | 'MA'
      | 'MT'
      | 'MS'
      | 'MG'
      | 'PA'
      | 'PB'
      | 'PR'
      | 'PE'
      | 'PI'
      | 'RJ'
      | 'RN'
      | 'RS'
      | 'RO'
      | 'RR'
      | 'SC'
      | 'SP'
      | 'SE'
      | 'TO';
    verified?: boolean | null;
  };
  cpf?: string | null;
  phoneNumber?: string | null;
  about?: string | null;
  cities?: (string | City)[] | null;
  states?:
    | (
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO'
      )[]
    | null;
  places?: (string | Place)[] | null;
  clinicalAreas?: (string | ClinicalArea)[] | null;
  graduationYear?: number | null;
  allRequiredInfoFilled?: boolean | null;
  moreRegisters?:
    | {
        description?: string | null;
        type: 'RQE';
        number: number;
        uf:
          | 'AC'
          | 'AL'
          | 'AP'
          | 'AM'
          | 'BA'
          | 'CE'
          | 'DF'
          | 'ES'
          | 'GO'
          | 'MA'
          | 'MT'
          | 'MS'
          | 'MG'
          | 'PA'
          | 'PB'
          | 'PR'
          | 'PE'
          | 'PI'
          | 'RJ'
          | 'RN'
          | 'RS'
          | 'RO'
          | 'RR'
          | 'SC'
          | 'SP'
          | 'SE'
          | 'TO';
        verified?: boolean | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  email: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  _verified?: boolean | null;
  _verificationToken?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "cities".
 */
export interface City {
  id: string;
  label: string;
  name: string;
  uf: string;
  keys: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "places".
 */
export interface Place {
  id: string;
  name: string;
  city: string | City;
  formattedAddress?: string | null;
  postcode?: number | null;
  /**
   * @minItems 2
   * @maxItems 2
   */
  coordinates?: [number, number] | null;
  rating?: number | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "clinical-areas".
 */
export interface ClinicalArea {
  id: string;
  name?: string | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface ProfileGroup {
  id: string;
  name: string;
  owner: string | Profile;
  matrixRoomId?: string | null;
  description?: string | null;
  picture?: (string | null) | Media;
  managers?: (string | Profile)[] | null;
  members?: (string | Profile)[] | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface JobModality {
  id: string;
  name?: string | null;
  clinicalAreas?: (string | ClinicalArea)[] | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "job-templates".
 */
export interface JobTemplate {
  id: string;
  from?: (string | null) | Profile;
  modality?: (string | null) | JobModality;
  clinicalArea?: (string | null) | ClinicalArea;
  state?:
    | (
        | 'AC'
        | 'AL'
        | 'AP'
        | 'AM'
        | 'BA'
        | 'CE'
        | 'DF'
        | 'ES'
        | 'GO'
        | 'MA'
        | 'MT'
        | 'MS'
        | 'MG'
        | 'PA'
        | 'PB'
        | 'PR'
        | 'PE'
        | 'PI'
        | 'RJ'
        | 'RN'
        | 'RS'
        | 'RO'
        | 'RR'
        | 'SC'
        | 'SP'
        | 'SE'
        | 'TO'
      )
    | null;
  city?: (string | null) | City;
  place?: (string | null) | Place;
  startTime?: string | null;
  durationInHours?: number | null;
  paymentMethod?: ('AV' | 'NR' | 'AC') | null;
  singlePaymentForMutipleDates?: boolean | null;
  description?: string | null;
  priceInCents?: number | null;
  restrictedToGuests?: (string | Profile)[] | null;
  restrictedToGroups?: (string | ProfileGroup)[] | null;
  visibility?: ('PUBLIC' | 'UNLISTED' | 'RESTRICTED_TO_GUESTS' | 'RESTRICTED_TO_GROUPS' | 'PRIVATE') | null;
  updatedAt: string;
  createdAt: string;
}